Theme as type + some collateral fixes
In the FP space, booleans are bad as they are not explicit. There exists
more problems with the booleans, namely that W3C does have future plans
to extend the light/dark to more themes which we wouldn’t be handling.
Also a boolean locks _us_ into a binary theme (what if we want an
OLED-friendly theme that is `#000`, etc.). The new `getTheme` function
properly runs through the problem space with the fallback. Along with
extending beyond two theme, the new structure can express unsetting an
explicit value as there is no way to do this currently. `isDarkTheme` is
still a useful variable around the app because the themes are tied to
material UI styles though I think this is a bug too and they shouldn’t
be arbitrarily limiting users either. Building our types around our
tools instead of our ideas is not a great way to handle things. Their
styles module seems deprecated in v5. Unsure, but maybe they saw this as
flawed as well.

Personally I think the switch toggle is bad as it can’t unselect _and_
is not future-proof—but this is the trend.

Collateral fixes::
* Some tests on the component side were failing, now fixed
* Some text-transform: uppercase
* Reverse some `!`’d ternaries which are confusing
* Remove useless `alt` tags, or made more clear