Implement Staging/Production HCI Effects
Using Hercules CI, I have created two functions which wrap hci-effects.
mkWebsite and mkGitBranchViaEffect. They are shell scripts which push to
branches on codeberg-pages. codeberg-pages then serves these branches,
which we then point to via Cloudflare.

This means that whenever code is merged into `staging`, each website is
pushed to a branch on codeberg-pages which is then served at the
following URLs:

- https://staging-frontend-dashboard-dc1ece41.ardana.org/
- https://staging-frontend-landing-dc1ece41.ardana.org/
- https://staging-frontend-vault-dc1ece41.ardana.org/

dc1ece41 is just an identifier to prevent robots or humans from
discovering the URL by brute force, it is not particularly important or
significant.